+++ packages/asset-check/lib/index.ts
@@ -3,4 +3,5 @@ export * from './cardano';
 export * from './bitcoin';
 export * from './evm';
 export * from './doge';
+export * from './firo';
 export * from './handshake';